+++ src/providers/ConfirmProvider/ConfirmProviderConstants.ts
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import { ConfirmFn } from "./ConfirmProviderTypes";
+
+export const ConfirmContext = createContext<ConfirmFn>(() => Promise.resolve(false));