+++ src/providers/index.tsx
@@ -1,11 +1,14 @@
 import { PropsWithChildren } from "react";
 import ContextMenuProvider from "./ContextMenuProvider";
 import ToastProvider from "./ToastProvider";
+import ConfirmProvider from "./ConfirmProvider";
 
 const GlobalProvider = ({ children }: PropsWithChildren) => (
   <ContextMenuProvider>
     <ToastProvider>
-      {children}
+      <ConfirmProvider>
+        {children}
+      </ConfirmProvider>
     </ToastProvider>
   </ContextMenuProvider>
 );