+++ src/stores/SettingsStore/components/SettingsModal/components/index.ts
@@ -1,6 +1,8 @@
 export { default as AccountSettings } from './AccountSettings';
 export type { AccountSettingsProps } from './AccountSettings';
+export { default as ArtifactHelperSettings } from './ArtifactHelperSettings';
 export { default as FavoritesOverview } from './FavoritesOverview';
+export { default as GeneralSettings } from './GeneralSettings';
 export { default as PendingChangesModal } from './PendingChangesModal';
 export { default as SettingsContent } from './SettingsContent';
 export { default as SettingsOption } from './SettingsOption';